Python - refactor transaction
