+++ python/python/tests/test_transaction.py
@@ -2,9 +2,11 @@ from datetime import datetime
 from typing import List, Union
 
 import pytest
-from pybushka.async_commands.cluster_commands import ClusterTransaction
-from pybushka.async_commands.core import BaseTransaction
-from pybushka.async_commands.standalone_commands import Transaction
+from pybushka.async_commands.transaction import (
+    BaseTransaction,
+    ClusterTransaction,
+    Transaction,
+)
 from pybushka.constants import OK, TResult
 from pybushka.protobuf.redis_request_pb2 import RequestType
 from pybushka.redis_client import RedisClient, RedisClusterClient, TRedisClient
